This commit allows us to install from private remote git repositories (e.g. `rip install git@github.com:jmhodges/murmur_hash.git`).
It does so by assuming that if the source path of the code we want to
install ends with `.git`, there is a git repo at the other
end. Obviously, we cannot be guaranteed that it is not a local repo,
so we check to see if the source path refers to a local file even if
it does not start with file://.

Weird ass edge cases can be seen a mile away, but this lets us at many
more use cases and later refinements can clear out the bugs.

Oh, and this doesn't quite let rip self-host but it would be
interesting to make that work.