Have `rip ruby` check for (and use) $RUBYBIN env variable.
This is useful when `rip ruby` is used to invoke long running daemons,
e.g. a Rails or Sinatra app.

RUBYBIN=/path/to/jruby rip ruby cheat_env cheat.rb

Rip should work with any version of Ruby that respects the $RUBYLIB env
variable, so this is not needed when running apps from within the
`active` env in an interactive shell.