Rip should download the dependencies for a gem, and properly construct the dependency graph in the ripenv.
I've taken the liberty of extracting the gem commands into their own
module, since they're needed in two classes now.

Most of the work here was figuring out the best way to establish the
relationship between gem packages and their dependencies, given the
existence of meta packages for each gem. I think, through a combination
of juggling parent packages, I've managed to ensure that dependencies
stay associated with their parent, even if there's a meta package in
the way. Hopefully this is useful for all packages, and not just gems.